requirements.txt
python-telegram-bot==21.5
yt-dlp==2025.10.22
moviepy==2.2.1
aiohttp==3.13.1
requests==2.32.5
python-dotenv==1.1.1
